Moved example file to the correct folder as requested
